Add files via upload
This is parks data filtered from the original in our repository. The goal was to use total parks acreage by ward to see if there was any association at all with crime count data in 2016. 

I filtered because there has to be an operational definition of "park".  In this case, there had to be a high enough number of each type to have a chance at being in any ward. I also eliminated entries that most people would not mentally classify as a park (e.g. a boat slip, community garden). I eliminated rows that had no category, or I couldn't find a definition (e.g. spray feature, wtf is that) The original idea was to use greenspace, but we don't have that. Here are the types of parks in the CSV. Basketball courts, tennis courts, community centers, combined football-soccer fields, junior baseball fields, gymnasium, 2 types of playgrounds and outdoor pools.

That's my operational definition of "park" for this exercise. If someone disagrees, and someone surely will, that's ok.