Add files via upload
My jupyter notebook for this parks crime stuff. I really wanted to do the poisson regression and so on but just quit caring at least for now. It's much faster and easier to do in R, SAS, SPSS and so on, but we'll only briefly talk about R in this class/camp/course whatever it is. I've found ways to run R in Python, none have worked yet, but I'll crack that eventually.